Security Scan Upload
Adds a workflow for weekly security scanning that
can upload to both artifact and gh security tab.

As some of our repos are private witohut licenses
we'll have to just uploads artifacts for now.

We left the option to turn sarif upload off
and gh security tab on only to hide results
in public repos / remove unncessary uploading.